+++ art-community-platform/frontend/src/components/Profile.css
