+++ art-community-platform/frontend/src/components/Profile.js
@@ -2,92 +2,24 @@ import {
   Col,
   Layout,
   Row,
-  Input,
   Space,
   Card,
   Pagination,
   Image,
-  Typography,
   Divider,
 } from "antd";
-import { useNavigate } from "react-router-dom";
 
 import React from "react";
-import { useDispatch, useSelector } from "react-redux";
-
-import { Logout as LogoutHelper } from "../utils/helper";
 
 import { UserOutlined } from "@ant-design/icons";
-import "./Profile.css";
-import Colors from "../constants/Colors";
-import { Homepage as HomepageHelper } from "../utils/helper";
-const { Header, Content, Footer } = Layout;
-const { Search } = Input;
-const { Text } = Typography;
+import Navbar from "./Navbar";
+const { Content, Footer } = Layout;
 
-const buttonStyle = {
-  fontSize: "20px",
-  color: Colors.primary,
-  cursor: "pointer",
-};
-
-const navbarStyle = {
-  color: Colors.primaryDark,
-  cursor: "pointer",
-};
 const Profile = () => {
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const { token } = useSelector((state) => state.login);
-
-  const sendToSignout = (values) => {
-    console.log("Success:", values);
-    LogoutHelper({}, dispatch);
-    navigate("/login");
-  };
-
-  const sendToHome = async (values) => {
-    console.log("Success to home:", token);
-    const response = await HomepageHelper({ token });
-
-    console.log("Success to home:", response);
-
-    navigate("/home");
-  };
-
-  const sendToProfile = (values) => {
-    console.log("Success:", values);
-
-    navigate("/profile");
-  };
 
   return (
     <Layout>
-      <Header style={navbarStyle}>
-        <Row>
-          <Col span={2}></Col>
-          <Col span={4}>
-            <Text style={buttonStyle} onClick={sendToHome}>
-              Home
-            </Text>
-          </Col>
-          <Col span={2}></Col>
-          <Col span={4}>
-            <Text style={buttonStyle} onClick={sendToProfile}>
-              Profile
-            </Text>
-          </Col>
-          <Col span={6}>
-            <Search placeholder="input search text" enterButton />{" "}
-          </Col>
-          <Col span={3}></Col>
-          <Col span={3}>
-            <Text style={buttonStyle} onClick={sendToSignout}>
-              Sign Out
-            </Text>
-          </Col>
-        </Row>
-      </Header>
+      <Navbar />
       <Content>
         <Row>
           <Col span={6}>
